refactor(: extract player menu sync into dedicated module and simplify state
